Moved PrintfCore logic into Print.cpp and optimized away some of the function calls
Saves another 200 bytes or so

+++ baremetal/PrintFloat.cpp
@@ -121,7 +121,7 @@ uint32_t Print::printString(Printer printer, Float v, int32_t precision, Capital
         n = -n;
     }
     
-    size += printString(printer, n);
+    size += printString(printer, static_cast<uint64_t>(n));
     printer('\0');
     return size;
 }
